Move priority_router related code to minecode/collectors #515
