+++ minecode/collectors/generic.py
