+++ packagedb/management/commands/fix_purl_values.py
@@ -21,8 +21,8 @@ import requests
 
 from minecode.management.commands import VerboseCommand
 from minecode.utils import MemorySavingQuerysetIterator
-from minecode.miners.maven import collect_links_from_text
-from minecode.miners.maven import filter_for_artifacts
+from minecode.collectors.maven import collect_links_from_text
+from minecode.collectors.maven import filter_for_artifacts
 from packagedb.models import Package
 
 DEFAULT_TIMEOUT = 30
